+++ src/Domain/Options/MariaDb/Field/JsonOptions.php
@@ -6,7 +6,7 @@ namespace SylvainDuval\DynamicDbBundle\Domain\Options\MariaDb\Field;
 
 use SylvainDuval\DynamicDbBundle\Domain\Options\FieldOptionsInterface;
 
-readonly class JsonOptions implements FieldOptionsInterface
+final readonly class JsonOptions implements FieldOptionsInterface
 {
 	/**
 	 * @param ?array<mixed> $default